# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * effimero. streetwear brand API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProductCategory =
  (typeof ProductCategory)[keyof typeof ProductCategory];

export const ProductCategory = {
  tshirts: "tshirts",
  pants: "pants",
} as const;

export interface Product {
  id: number;
  name: string;
  description: string;
  category: ProductCategory;
  imageUrl?: string | null;
  sizes: string[];
  featured: boolean;
  createdAt: string;
}

export type CreateProductRequestCategory =
  (typeof CreateProductRequestCategory)[keyof typeof CreateProductRequestCategory];

export const CreateProductRequestCategory = {
  tshirts: "tshirts",
  pants: "pants",
} as const;

export interface CreateProductRequest {
  name: string;
  description: string;
  category: CreateProductRequestCategory;
  imageUrl?: string | null;
  sizes: string[];
  featured?: boolean;
}

export interface CreateInquiryRequest {
  productId: number;
  name: string;
  email: string;
  message?: string | null;
  size?: string | null;
}

export interface Inquiry {
  id: number;
  productId: number;
  name: string;
  email: string;
  message?: string | null;
  size?: string | null;
  createdAt: string;
}

export interface SiteSettings {
  wipEnabled: boolean;
  wipCountdown?: string | null;
  wipMessage?: string | null;
}

export interface UploadResponse {
  url: string;
  filename: string;
}

export interface ErrorResponse {
  error: string;
}

export type ListProductsParams = {
  category?: ListProductsCategory;
  featured?: boolean;
};

export type ListProductsCategory =
  (typeof ListProductsCategory)[keyof typeof ListProductsCategory];

export const ListProductsCategory = {
  tshirts: "tshirts",
  pants: "pants",
  all: "all",
} as const;

export type UploadImageBody = {
  file: Blob;
};