# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: effimero. streetwear brand API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product catalog operations
  - name: settings
    description: Site settings
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [tshirts, pants, all]
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /inquiries:
    post:
      operationId: createInquiry
      tags: [products]
      summary: Submit a product inquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInquiryRequest"
      responses:
        "201":
          description: Inquiry submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inquiry"
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get site settings
      responses:
        "200":
          description: Site settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"
    put:
      operationId: updateSettings
      tags: [settings]
      summary: Update site settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteSettings"
      responses:
        "200":
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"
  /upload:
    post:
      operationId: uploadImage
      tags: [products]
      summary: Upload an image
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        "200":
          description: Upload successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [tshirts, pants]
        imageUrl:
          type: string
          nullable: true
        sizes:
          type: array
          items:
            type: string
        featured:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - category
        - sizes
        - featured
        - createdAt
    CreateProductRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [tshirts, pants]
        imageUrl:
          type: string
          nullable: true
        sizes:
          type: array
          items:
            type: string
        featured:
          type: boolean
      required:
        - name
        - description
        - category
        - sizes
    CreateInquiryRequest:
      type: object
      properties:
        productId:
          type: integer
        name:
          type: string
        email:
          type: string
        message:
          type: string
          nullable: true
        size:
          type: string
          nullable: true
      required:
        - productId
        - name
        - email
    Inquiry:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        name:
          type: string
        email:
          type: string
        message:
          type: string
          nullable: true
        size:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - productId
        - name
        - email
        - createdAt
    SiteSettings:
      type: object
      properties:
        wipEnabled:
          type: boolean
        wipCountdown:
          type: string
          nullable: true
        wipMessage:
          type: string
          nullable: true
      required:
        - wipEnabled
    UploadResponse:
      type: object
      properties:
        url:
          type: string
        filename:
          type: string
      required:
        - url
        - filename
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error