# lib/api-zod/src/generated/types/siteSettings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * effimero. streetwear brand API
 * OpenAPI spec version: 0.1.0
 */

export interface SiteSettings {
  wipEnabled: boolean;
  wipCountdown?: string | null;
  wipMessage?: string | null;
}